requirements.txt
discord.py==2.3.2
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
mercadopago==2.2.1
python-dotenv==1.0.1
cryptography==42.0.8
httpx==0.27.0
